+++ changelog.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## UNRELEASED
+## v31.1
 
 ### Features